feat: refactor networks
