+++ mava/configs/network/cnn.yaml
@@ -1,7 +1,7 @@
 # ---CNN Networks---
 actor_network:
   pre_torso:
-    _target_: mava.networks.CNNTorso
+    _target_: mava.networks.torsos.CNNTorso
     channel_sizes: [32, 32]
     kernel_sizes: [3, 3]
     strides: [1, 1]
@@ -9,11 +9,11 @@ actor_network:
     activation: relu
 
 action_head:
-  _target_: mava.networks.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
+  _target_: mava.networks.heads.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
 
 critic_network:
   pre_torso:
-    _target_: mava.networks.CNNTorso
+    _target_: mava.networks.torsos.CNNTorso
     channel_sizes: [32, 32]
     kernel_sizes: [3, 3]
     strides: [1, 1]
