+++ mava/configs/network/rcnn.yaml
@@ -3,31 +3,31 @@ hidden_state_dim: 128 # The size of the RNN hiddenstate for each agent.
 
 actor_network:
   pre_torso:
-    _target_: mava.networks.CNNTorso
+    _target_: mava.networks.torsos.CNNTorso
     channel_sizes: [32, 32]
     kernel_sizes: [3, 3]
     strides: [1, 1]
     use_layer_norm: False
     activation: relu
   post_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
 
 action_head:
-  _target_: mava.networks.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
+  _target_: mava.networks.heads.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
 
 critic_network:
   pre_torso:
-    _target_: mava.networks.CNNTorso
+    _target_: mava.networks.torsos.CNNTorso
     channel_sizes: [32, 32]
     kernel_sizes: [3, 3]
     strides: [1, 1]
     use_layer_norm: False
     activation: relu
   post_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
