+++ mava/configs/network/rnn.yaml
@@ -3,39 +3,39 @@ hidden_state_dim: 128 # The size of the RNN hiddenstate for each agent.
 
 actor_network:
   pre_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
   post_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
 
 action_head:
-  _target_: mava.networks.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
+  _target_: mava.networks.heads.DiscreteActionHead # [DiscreteActionHead, ContinuousActionHead]
 
 critic_network:
   pre_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
   post_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
 
 q_network:
   pre_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
   post_torso:
-    _target_: mava.networks.MLPTorso
+    _target_: mava.networks.torsos.MLPTorso
     layer_sizes: [128]
     use_layer_norm: False
     activation: relu
